+++ src/nvlib/view/properties_window/character_view.py
@@ -138,7 +138,10 @@ class CharacterView(WorldElementView):
                 date.fromisoformat(birthDateStr)
             except:
                 self._birthDate.set(self._element.birthDate)
-                self._ui.show_error(f'{_("Wrong date")}: "{birthDateStr}"', title=_('Input rejected'))
+                self._ui.show_error(
+                    f'{_("Wrong date")}: "{birthDateStr}"\n {_("Required")}: {_("YYYY-MM-DD")}',
+                    title=_('Input rejected')
+                    )
             else:
                 self._element.birthDate = birthDateStr
 
@@ -150,7 +153,10 @@ class CharacterView(WorldElementView):
                 date.fromisoformat(deathDateStr)
             except:
                 self._deathDate.set(self._element.deathDate)
-                self._ui.show_error(f'{_("Wrong date")}: "{deathDateStr}"', title=_('Input rejected'))
+                self._ui.show_error(
+                    f'{_("Wrong date")}: "{deathDateStr}"\n {_("Required")}: {_("YYYY-MM-DD")}',
+                    title=_('Input rejected')
+                    )
             else:
                 self._element.deathDate = deathDateStr